数据缓存1.0

+++ app/src/main/java/com/example/emailmanager/emaildetail/EmailDetailActivity.java
@@ -3,24 +3,12 @@ package com.example.emailmanager.emaildetail;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
-import android.webkit.WebView;
-import android.widget.FrameLayout;
 
-import com.example.emailmanager.EMApplication;
 import com.example.emailmanager.R;
-import com.example.emailmanager.data.AccountDetail;
-import com.example.emailmanager.data.AccountDetailDao;
-import com.example.emailmanager.data.Email;
-import com.example.emailmanager.data.EmailDao;
 import com.example.emailmanager.data.source.EmailRepository;
 import com.example.emailmanager.databinding.ActivityEmailDetailBinding;
 import com.example.emailmanager.emaildetail.adapter.AccessoryListAdapter;
-
-import org.greenrobot.greendao.query.QueryBuilder;
-
-import java.util.List;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -63,7 +51,7 @@ public class EmailDetailActivity extends AppCompatActivity {
         });
     }
 
-    public static void start2EmailDetailActivity(Context context, int msgNum) {
+    public static void start2EmailDetailActivity(Context context, Long msgNum) {
         context.startActivity(new Intent(context, EmailDetailActivity.class).putExtra("msgnum", msgNum));
     }
 }
